+++ src/components/pages/apps/matcalc/history/history-provider/actions/action-types.ts
@@ -1,0 +1,48 @@
+import { ITextFileOpen } from '@/components/pages/open-files'
+import { IDBEntity } from '@/interfaces/db-entity'
+import { IClusterGroup, IClusterGroupRow } from '@/lib/cluster-group'
+import { type BaseDataFrame } from '@/lib/dataframe/base-dataframe'
+import type { IGeneSet } from '@/lib/gsea/geneset'
+import {
+  AppendMode,
+  FilePath,
+  HistoryPath,
+  HistoryPlot,
+  IGroupOps,
+  ISheetOps,
+} from '../history-types'
+
+export type HistoryAction =
+  | { type: 'reset' }
+  | { type: 'init' }
+  | { type: 'undo' }
+  | { type: 'redo' }
+  | { type: 'seek'; step: number | string }
+  | {
+      type: 'openFile'
+      file: IDBEntity
+      sheets: BaseDataFrame[]
+      plots: HistoryPlot[]
+      groupRows: IClusterGroupRow[]
+      genesets: IGeneSet[]
+      //groupsName: string
+      mode: AppendMode
+    }
+  | { type: 'addSheets'; sheets: BaseDataFrame[]; opts: ISheetOps }
+  | { type: 'addPlots'; plots: HistoryPlot[]; opts: ISheetOps }
+  | { type: 'remove'; paths: HistoryPath[] }
+  | { type: 'removeFiles'; paths: FilePath[] }
+  | { type: 'reorderSheets'; sheets: BaseDataFrame[]; opts: ISheetOps }
+  //| { type: 'reorderPlots'; ids: string[]; opts: ISheetOps }
+  | { type: 'updatePlot'; plot: HistoryPlot; opts: ISheetOps }
+  | { type: 'addGroups'; groupRows: IClusterGroupRow[]; opts: IGroupOps }
+  | { type: 'clearGroups'; opts: IGroupOps }
+  | { type: 'updateGroup'; group: IClusterGroup; opts: IGroupOps }
+  | { type: 'removeGroups'; ids: string[]; opts: IGroupOps }
+  | { type: 'openGroupFiles'; files: ITextFileOpen[]; opts: IGroupOps }
+  | { type: 'addGenesets'; genesets: IGeneSet[]; opts: IGroupOps }
+  | { type: 'clearGenesets'; opts: IGroupOps }
+  | { type: 'updateGeneset'; geneset: IGeneSet; opts: IGroupOps }
+  | { type: 'removeGenesets'; ids: string[]; opts: IGroupOps }
+  //| { type: 'reorderGenesets'; ids: string[]; opts: IGroupOps }
+  | { type: 'goto'; path: HistoryPath }