+++ src/components/pages/apps/matcalc/history/history-provider/history-manager.ts
@@ -30,8 +30,9 @@ export interface IUndoState<T> {
   history: IHistoryEntry<T>[]
   cursor: number
   /**
-   * The current version of the history state which is a monotonically increasing number
-   * and can be used to detect changes in the history state.
+   * The current version of the history state which is a monotonically
+   * increasing number and can be used to detect changes in the
+   * history state when other properties do not indicate change.
    */
   version: number
 }